+++ kite-core/src/main/kotlin/com/tang/kite/annotation/CompositeIndex.kt
@@ -19,7 +19,7 @@ annotation class CompositeIndex(
     /**
      * Whether the index in unique or not
      */
-    val unique: Boolean,
+    val unique: Boolean = false,
 
     /**
      * Index type