+++ kite-core/src/main/kotlin/com/tang/kite/sql/ast/SqlNode.kt
@@ -186,7 +186,7 @@ sealed class SqlNode {
 
     data class DropIndex(
 
-        var indexName: String? = null,
+        var indexName: String,
 
         var table: TableReference? = null