+++ kite-core/src/test/kotlin/com/tang/kite/schema/synchronization/index/dropindex/AccountDropIndex.kt
@@ -1,7 +1,6 @@
-package com.tang.kite.schema.synchronization.index.addindex
+package com.tang.kite.schema.synchronization.index.dropindex
 
 import com.tang.kite.annotation.CompositeIndex
-import com.tang.kite.annotation.Index
 import com.tang.kite.annotation.Table
 import com.tang.kite.annotation.field.CreateTime
 import com.tang.kite.annotation.field.UpdateTime
@@ -10,18 +9,17 @@ import com.tang.kite.annotation.id.IdType
 import java.time.LocalDateTime
 
 /**
- * Test entity with fewer columns for schema synchronization test (add index scenario)
+ * Test entity with fewer columns for schema synchronization test (drop index scenario)
  *
  * @author Tang
  */
 @CompositeIndex(unique = true, columns = ["username", "nickname"], orders = ["asc", "desc"])
-@Table(value = "account", comment = "Table for account add index")
-class AccountAddIndex(
+@Table(value = "account", comment = "Table for account drop index")
+class AccountDropIndex(
 
     @Id(type = IdType.AUTO)
     var id: Long? = null,
 
-    @Index
     var username: String? = null,
 
     var nickname: String? = null,