+++ kite-core/src/test/kotlin/com/tang/kite/schema/synchronization/index/modifyindex/AccountModifyIndex.kt
@@ -1,4 +1,4 @@
-package com.tang.kite.schema.synchronization.index.addindex
+package com.tang.kite.schema.synchronization.index.modifyindex
 
 import com.tang.kite.annotation.CompositeIndex
 import com.tang.kite.annotation.Index
@@ -10,18 +10,18 @@ import com.tang.kite.annotation.id.IdType
 import java.time.LocalDateTime
 
 /**
- * Test entity with fewer columns for schema synchronization test (add index scenario)
+ * Test entity with fewer columns for schema synchronization test (modify index scenario)
  *
  * @author Tang
  */
 @CompositeIndex(unique = true, columns = ["username", "nickname"], orders = ["asc", "desc"])
-@Table(value = "account", comment = "Table for account add index")
-class AccountAddIndex(
+@Table(value = "account", comment = "Table for account modify index")
+class AccountModifyIndex(
 
     @Id(type = IdType.AUTO)
     var id: Long? = null,
 
-    @Index
+    @Index(order = "desc")
     var username: String? = null,
 
     var nickname: String? = null,